Add privacy feature to adnslogres
The option -P prints of successfully looked up DNS names in the format

  phhhhhhhh.t

where P is the literal character 'p', H is a hex digit and T is the
top level domain name.  The hex digits are the truncated hash value of
the entire domain name.  To later lookups impossible the option -S
allows to gives a salt value for this run.