src/: Separate context state into parts for the type itself and its parent.
The `qcontext.info' union has two distinct uses.  A PTR query uses it to
store the initial query address, against which it compares the answer to
its child A query.  On the other hand, `pap_hostaddr' uses this space in
each child query it makes, to keep track of where to put the socket
addresses when the answers come back.

These two kinds of uses are incompatible, so split `qcontext.info' into
two pieces: one for the internal use of a query type, and one for the
use of a child query's parent -- and specifically for its completion
callback.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>