src/types.c: Rearrange the main typeinfo table.
Kill XTRA_TYPE, which won't scale at all well as new features are added
to the typeinfo structure.  Instead, allow additional arguments to
FLAT/DEEP_TYPE, which are used to carry designated initializers for the
more unusual fields.

I took the opportunity to factor out some common prefixes in the names
of the functions listed in the table, which gives us some additional
horizontal space.  This will come in useful later.

All of this involves a little nontrivial macrology, now in internal.h.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>