src/: adns__internal_submit can modify ctx->tinfo.
It will be convenient for type hooks to store information in their part
of the query context before the query is properly established.  We could
have adns__internal_submit make a copy of the caller's context, but it
turns out that none of them actually care if we change it.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>