src/, regress/: Prepare for early failures in PTR queries.
Soon we're going to change the way adns_r_ptr queries work: in
particular, queries with names which can't be parsed back into addresses
will be rejected during submission rather than waiting for the answer to
the initial PTR question to come back.

So here we apply a really simple temporary hack to adns_submit which
approximately captures the behaviour we're after, and adjust the various
test cases to match.  We'll take this hack out when we make the proper
change to the PTR handling.

Hopefully this will persuade reviewers that there's nothing up my
sleeves while I mess with the regression tests.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>