Add macro to safely clear memory.
* src/internal.h (WIPEMEMORY): New.
--

This kind of platform neutral code has been in use by GnuPG and
Libgcrypt for ages.  I am still waiting for some C committee f^D
experts to figure that this makes use of undefined behaviour for
volatile and they tell their optimizing-for-the-flat-world compiler
to remove such code and thereby unveil passwords in memory (which
actually happened for the standard memset).

Signed-off-by: Werner Koch <wk@gnupg.org>