Catch network failures from sendto earlier.
* src/adns.h (adns_s_netunreach): New.
(adns_s_netdown): New.
* src/general.c (struct sinfo): Add corresponding strings.
* src/transmit.c (adns__query_send): Bail out early on ENETUNREACH and
ENETDOWN.
--

This patch will stop retrying sending UDP packets to an IP address for
which the sever can't reach the network or the interface is down.  A
common case for these errors are lost network connections (WLAN lost
or cable unplugged).  It does not make sense to keep on trying once
sendto returned one of these errors - it is unlikely that this will be
solved within the next 30 seconds (15 retries * 2s)

Signed-off-by: Werner Koch <wk@gnupg.org>